Fix a problem with loading and saving files, specifically those with metadata:
If there is an error opening the file to be saved for write,
`on-save-file' is called, but `after-save-file' isn't.  As
`on-save-file' creates the metadata, and `after-save-file' zaps it
again, corruption results.  To fix, open the file inside the relevant
`dynamic-wind' rather than before.

The analogous problem existed with loading.

svn: r7602

original commit: 3ab3a7300c50504526030cb4e5bbdfd5e5e922a2