Update the text message at the top of WXME files, including the URL.
Also, update the message in existing files.  Change the instructions to
use DrRacket or DrScheme version <previous-version>, and for really old
files (pre-v4), just use 4 for the version.  Also just drop the second
paragraph in these files -- "Most likely, it was created by" is not
needed when for these files we know that this is true.

original commit: bba676d90ca86777b94e9a8264907a22e36bfea3