win32: fix canvas refresh
Refresh was wrong in the case that a canvas had been "reset"
in certain ways, such as showing and hiding, and the canvas is
drawn on before a Win32-level refresh event was processed. In
that case `on-paint' wasn't called, and it should have been.

Closes PR 12152

original commit: f5e534fd87e1e4e278e835c9237c2efa5ed27bc2