ffi/unsafe/objc: remove `-A'/`+A', add `#:async-apply'
Another run at the problem 57516164de was meant to solve. The
new solution just gives up when a `drawRect:' method is called
in the wrong thread, which might create refresh glitches but
avoids a deadlock in the case that the Racket thread is blocked
on the update happening in the foreign thread.

original commit: ca0418d47d62426a4321ddfe7001466f7d87739e