racket/gui cocoa: repair for canvas creation in hidden top-level window
Fixes 9cb646bbd2 for the case that a frame or dialog is shown,
then hidden, then a new canvas is created or a new compatible
bitmap is requested via the `make-bitmap` method.

The test for this change is just the "windowing.rktl" test, which I
forgot to run before (as opposed to the drawing test suite).

original commit: d234c0c2e30e7ed0c649f8bcf2577e23865304c3