Remove dependence on libgtkgl; fix indirect rendering problem
The right way to render to pixmaps is to create a GLX pixmap wrapper
and render to *that*. Almost nobody does this - including libgtkgl -
and it's almost never a problem. But it causes crashes on my system
in indirect rendering mode.

This commit changes three things.

 1. OpenGL on Linux no longer requires libgtkgl, only libGL, which
    comes preinstalled on many (most? almost all?) systems.

 2. Rendering to pixmaps is done properly, via a GLX pixmap wrapper.

 3. Direct rendering is done whenever possible, even for pixmaps.

original commit: c601b82f04586b686583107b80aceef4e8606ae1