fix drawing bug in line-numbers and special-first-line support
This drawing bug was that the DC's state wasn't being
initialized all of the way. In particular, the text-mode
wasn't being set and so, in rare occasions, other drawing
code would (legally) leave it in 'solid mode with a non-white
background color also set. Which would cause those two things
to draw incorrectly.