avoid using separate events to send keystrokes in a hope of avoiding intermittent failures in these tests
The guess that motivates this commit about what's going wrong is that
under some conditions that have to do with parallel tests that might
grab the OS-level focus, racket/gui is deciding that some of the two
keystroke combinations count as separate keystrokes, instead of being
combined in the search for the keymap handler. For example, if the
test suite tries to type c:c and then c:l, that will come in as two
separate events. If unlucky timing coincidence happens, then
racket/gui will decide that those should not count together in the
search for a keymap entry, but count as two separate ones (so we'd get
c:l centering the view instead of c:c;c:l inserting let-ans).