Fix bug where DrRacket could loose data when saving.
At the moment if a user refuses to save a file with non-textual widgets
using wxme, DrRacket will just save the file as flattened text. This leads
to lost data, and makes otherwise valid racket files invalid.

This patch prevents the file from loosing that data on saving. Namely,
if a user:
1. Opened a file as a text file.
2. Added non-textual snips to the file.
3. Has in their preferences to warn before changing the file type.
4. Saves the file but refuses to change the type.

then DrRacket will not save the file because it cannot
do so.