Remove call to gdk_window_process_all_updates
From https://developer.gnome.org/gdk3/stable/gdk3-Windows.html#gdk-window-process-all-updates

> gdk_window_process_all_updates has been deprecated since version 3.22 and should not be used in newly-written code.

Calls to this function seem to add significant overhead when, for example, drawing text in a simple text editor. There seems to be no benefit to this call, as removing it still will have things be redrawn on flush-display via gdk_display_flush. I am not sure if there are legacy reasons for the call to gdk_window_process_all_updates, but it would be great to not have this call if it is not necessary.