fix abuse of `_NSString`
Using `_NSString` to access global strings from Cocoa is wrong,
because the `_NSString` reference can get GCed and cause the
underlying Objective-C object to get released.

The problem manifests on my machine as mouse-wheel scrolling in a list
box sometimes freezing the GUI, and also the problem mentioned in the
deleted comment.