try again on display repair for Mojave
The previous attempts did not work right. The problem is that Apple
has removed frame-specific suspend of flushes, apparently in an effort
to simplify the API; as usual, though, it moves in a direction away
from multiple GUI contexts within an application. So, we adopt the
usual strategy of trying to make `racket/gui` work right in the common
case (here: suspends on the currently frontmost frame with no mouse or
keyboard activity on other frames) and have the behavior degrade as
gracefully as possible in the general case.