fix Racket Meta bindings for MacOS "Treat alt key as meta" mode
Fix Racket mode editor bindings to work right with the Option key as
Meta. There are two parts to the repair:

 * The `racket:text-mixin` and `racket:text-mode-mixin` keymap setup
   needs to use `#:alt-as-meta-keymap` and chain to the alt-as-meta
   keymap (when enabled).

   That way, the intent of M-C-<key> sets up a "?:c:a:<key>" mapping,
   for example.

 * The `keymap%` scoring mechanism should penalize "?:..." matches
   only when the "other" key mapping is needed to actually match.

   This adjustment helps because the Ctl- in Ctl-Option-<key> on MacOS
   prevents the Option- from modifying <key>. When "?" is always
   penalized, then "?:c:a:f" counts lower than "c:f" than
   Ctl-Option-f, for example, but it should count as a strictly better
   match.

There's a danger that the scoring change will have an undesirable
effect on matching for cases other than Ctl-Option-... on MacOS,
especially AltGr combinations on Windows. I think it's about as likely
to improve things in those cases, though. I don't have a better idea
than to just try it and solicit feedback for a snapshot build.