fix a shortcut in `find-position-in-line`
Internally, a `find-position-in-snip` variant is used to find
positions in a line, and the line search checks a snip whose right
edge matches the given location. A shortcut within
`find-position-in-snip` for a single-item snip expected to the
location to be strictly to the left of the right edge.

Also, fix the docs for `find-location`, which were a little
confused on the `edge-close` result.

Fixes racket/drracket#345