add-coloring and add-coloring/spell simplification
Since add-coloring/spell with false first argument simply passes the rest of the arguments to add-coloring, and the first argument is never dynamically false, inline those add-coloring calls and remove the parameter. That leaves only one use, so move out the single add-coloring call for uniformity and separation of concerns.

Now add-coloring/spell is simply responsible for registering region suggestions, so rename it to add-suggestions.

Also, move the “set if not set” of misspelled-regions to immediately before the use it's guarding.

Finally, name the parts of spell-info being used, for clarity and repeated use. To maximize potential acceptibility of this pull request, that change does not introduce any calls that weren't already happening, hence the asymmetry in accessing the first element, and doesn't convert the case expression to a more idiomatic cond. Those changes could increase the time for dispatch to a fast path, although that optimization seems trivial relative to the rest of the path.