Fix spell checking
New approach that Identifies spelling suggestions for clients by rechecking the token at the requested position, rather than trying to track shifting regions created when misspellings are colored.

The former implementation was buggy when misspelled regions changed token type or were edited without completely being replaced. A correct implementation following its approach would still be tricky, and prone to efficiency costs that the new approach shifts completely to when the user requests the location of a misspelled word or its suggestions.

The user now only pays for spell check features (find misspelled word, request suggestions) when used (and which is rare compared to general editing and recoloring). Moreover, the overhead is negligible relative to the user's interaction time with the feature (context switch, menu selection or keypress, consideration of and reaction to the result).