hierlist: fix outline selection drawing
In outline-highlighting mode, which is used when a hiehrlist doesn't
have the focus, a selected item's outline could be moved inside the
right and bottom edge of a display area. That wasn't consistent with
how the top and left are handled, but more importantly, it doesn't
play nicely with scrolling by copying. Fix the right and bottom edge
to always use the item's size.