cocoa: refresh a canvas even if not shown
The `refresh` method on a `canvas%` object was effectively ignored if
the current frame is no visible, but because the Mac OS windowing
system can hide or show windows, another refresh was not always
triggered on showing. To avoid that problem, just refresh a canvas's
content even if it's not shown.