the calls to size-allocate seem to be off by one for some unknown reason, but perhaps also not needed, so let's try disabling them
This is motivated by racket/drracket#350

It seems that, when there is a syntax error in the buffer and check
syntax's results come back, we get into a mode where the backing
bitmap for the definitions window is destroyed and recreated each time
the results come back. The off-by-one observed from these callbacks
triggers the destruction and the recreation of the backing bitmaps
each time.