adjust wxme save format in the interest of efficiency
This change abadons an earlier principle of the wxme format,
namely that just pasting the file into an email buffer would
not damage it, even if (say) all the 8th bits were zero'd.

This upside is that the decoding algorithm is much simpler (just
reads the bytes directly) and this translates into both smaller
sized files and faster loading time. The encoding isn't a lot
simpler but that speeds up too since most of the work is now
in write-bytes.

On an example that consists of a pict with four 400x400 bitmaps
embedded in it, using drracket's internal pict-snip, writing is about
25% faster and reading is about 5x faster.

This introduces version 10 of the wxme format