suprress "wrong eventspace for constrained event handling"
This logged error has appeared once in a while on Mac OS. It turns out
that `(tell evt window)` is approximate for keyboard and mouse events.
Even though the result does not seem to change, the actual window to
receive the event depends on the focus owner by the time the event is
dispatched, not when it is dequeued.

The program below illustrates. Assuming keyboard focus is enabled for
buttons, hit the space bar a couple of times, where the first one will
show an inner window, and the second one will close it. The `(sleep
0.1)` provides time for both the key-down and key-up for the second
space bar to be queued ostensibly for the inner window --- but when
the inner window goes away, the key-up event is actually dispatched to
the outer one. The result is that the key-up event is more or less
discarded, but also triggers error logging (previously).

While it may seem that dispatchig directly to the originally expected
window could solve this problem, that has some bad effects, such as
disabling menu shortcuts. Since there does not appear anything that
can be done about the dispatch in the wrong eventspace, just ignore
it and continue quietly.

 (define f (new frame% [label "Outer"]))
 (new button%
      [parent f]
      [label "Inner"]
      [callback (lambda (b e)
                  (parameterize ([current-eventspace (make-eventspace)])
                    (queue-callback
                     (lambda ()
                       (define g (new frame% [label "Inner"]))
                       (new button%
                            [parent g]
                            [label "Close"]
                            [callback (lambda (b e)
                                        (sleep 0.1)
                                        (send g show #f))])
                       (send g show #t)))))])
 (send f show #t)