gtk: fix oversampling for backing bitmap
When creating a backing bitmap for a canvas, the bitmap scaling was
forced to be an integer factor (accidentally, I think). As a result,
when copying the bitmap to the screen with a non-integer scaling
factor, the copy required scaling, which can make refresh much slower.
This commit avoids rounding up to an integer for the bitmap's backing
scale.