move default indentation and expression navigation to a separate library
Extract the implementation of S-expression indentation and navigation
(forward, backward, up, and down) to a set of functions outside of
`racket:text-mixin`. That way, the implementation can be used in other
editing environments.

The `color:text<%>` interface now also inherits from
`color-textoid<%>` from the "syntax-color" package. That interface
defines methods that can be used for indentation and navigation
purposes without `racket/gui`.