windows: revise transparent canvas refresh
This commit mostly reverts a34d1698, and instead addresses the problem
with flickering by moving the timing of the erase step. The previous
approach didn't actually use a non-transparent backing store due to
the use of `get-canvas-background` in `make-backing-bitmap`. Another
key change is an extra callback to `on-backing-flush` when there's
nothing to draw.