windows transparent canvas refresh, take 3
Revises b406ddfc28, which revised 2818e90be3. The strategy here goes
back to the intent of 2818e90be3 in that it uses a backing bitmap that
does not have an alpha channel, but it's simpler and avoids mistakes
in that earlier attempt. This revision also keeps an improvement in
b406ddfc28 to avoid unecessary an clear just before a successful
refresh. A specific goal here is to reduce flicker for the tab-panel
control in DrRacket.