add grapheme support to the editor classes
The `snip%` class gets three new methods: `get-grapheme-count`,
`grapheme-position` (converts from graphemes to characters), and
`position-grapheme` (covert from characters to graphemes). The default
implementares the same as `get-count` and identity functions,
respectively, but they're specialized for `string-snip%`.

The `text%` class also gets `grapheme-position` and
`position-grapheme`, plus some grapheme sensitively (automatic for
characters, on demand for strings) in `insert` and (automatic) in
`move-position`, `delete`, and `on-default-char`. For compatibility,
however, items and positions still corerspond to characters.