revised grapheme approach: grapheme positions by default
Commit ee68056a75 added support for graphemes and grapheme positions,
but left all existing methods as using char-based positions. That
approach was based on the fact that many things are deeply tied to
characters instead of graphemes, especially internally, and so isn't
practical to make grapheme counting the only interface.

This commit shifts to grapheme counting by default, which is enabled
by keeping the old methods with a `/char` suffix while adding new
versions of the old methods to perform grapheme<->char converions on
entry and exit. This is not a backward-compatible change, because it
shifts the relationship of existing methods and characters. It's a
more forward-compatible: it makes most existing code work the right
way with multi-character graphemes, and implicitly adapts an interface
like `color-textoid<%>` to support multi-character graphemes. For the
rare case that this change breaks a grogram, the program could be
changed to use `/char` methods.

The `snip%` interface is more "internal" than "external", and it
sticks a char-based interface, continuing to rely on explicit support
for grapheme<->char conversion.