Better syncronizaiton in the autosave tests
I guess it might be possible to access to the frontmost frame before
it has been totally filled in, which could make `wait-for-recover` return
but the button-pushing function not actually find the button.

If that's what's going wrong with these tests, then this change'll
make it so that we don't move on until the frame is fully filled in.