Update gtk to react to dark mode changes
The change detects entering and exiting dark mode by looking for
"-dark" in the name of the theme when themes switch. I wasn't able to
find anything more explicit than that.

When switching, the code allocates a new value and saves it to
`the-text-style` -- this causes
`get-label-{foreground,background}-color` to start reporting different
results. It also uses the referesh-all-canvases codepath (introduced
for the Mac OS dark mode changes) and inokes the
application-dark-mode-handler