Disable asynchronous mode for GTK printing (#336).
Fix issue #336: for unknown reasons, asynchronous printing in GTK3 generates
blank pages.  Disabling asynchronous mode fixes this problem.  This change makes
the main window freeze when the print dialog is popped up.  But it shouldn't be
a problem.  First this is also the default behavior when the print dialog is
popped up in Windows.  This change makes the behavior more consistent on
different OSes.  Besides, it is reasonable to forbid editing when the contents
of the editor is being printed.