Mac: repair label cleaning to remove all "&" escapes, not just the first
A button label like "U&pdate" means that "P" is a keyboard shortcut
for the button. A button label like "U&pda&te" doesn't really make
sense, but Windows and Unix render that as "Update", while Mac OS
rendered it as "Upda&te". The change here makes Mac OS consistent with
the others, which is also consistent with the documentation and with
`gui-utils:quote-literal-label`.