Adjust the way dark mode works and its interaction with colorschemes
In particular, we no longer keep a preference that is "the current
color scheme".  Instead, we keep two preferences with a slightly
different meaning: "the name of the preferred dark-mode color scheme"
and "the name of the preferred light-mode color scheme". This subtle
shift in the meaning of color schemes cleans up a number of things in
the code. The "current color scheme" is now an emergent property of
what the colors happen to be.

We also add a new preference that tells us if we're paying attention
to the underlying OS's dark mode setting or not (and, if we're not, do
we want to be in dark mode or light mode).

Also, the color preferences dialog now has checkboxes saying which
color scheme to use in dark mode and which one to use in light mode
instead of buttons that set the color scheme. Choosing the chcekboxes
might still trigger a change in the colorscheme (if we're in the
corresponding mode)

As part of this, code should be using
color-prefs:white-on-black-color-scheme? to determine if it should
draw in dark or light mode now instead of other ways. Also, this
points to a future improvement: there is code that shouldn't depend on
the framework that might want to pay attention to dark or light mode,
so the color schemes should probably move into mrlib/ somewhere,
someday.

related to racket/drracket#469